Reduce polution of the global namespace.
Undeclared variables pollute the global namespace and are annoying to track down. This change adds var in-front of x in the for loop and the body variable preventing them from being attached to the global object. A long term solution to making sure a problem like this does not happen in the future could be adding 
```
"use strict";
```
to the top of the project.