Fix invalid escape sequences from `buffer.toString()`
There seems to be some form of issue when converting the response buffer into a string. The JSON coming from Cleverbot is properly escaped, but converting a buffer to JSON is harder than I expected. I need to do more research.

For the time being, I am putting in a dirty `eval` to try to fix it. `eval`ing the response from Cleverbot fixes the formatting.